VisQA+/requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
cmasher==1.6.3
colorspacious==1.1.2
cycler==0.11.0
demjson==2.2.4
e13tools==0.9.6
fastdtw==0.3.4
flake8==4.0.1
fonttools==4.28.4
imageio==2.13.5
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.3.3
mccabe==0.6.1
memory-profiler==0.60.0
natsort==8.0.2
networkx==2.6.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
psutil==5.9.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
regex==2021.11.10
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Shapely==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tqdm==4.62.3
beautifulsoup4==4.10.0
